rename conversation to messages
