+++ src/routes/index.js
@@ -6,7 +6,7 @@ import App from '../App';
 import Home from '../pages/Home/Home';
 import Login from '../pages/Login/Login';
 import Shared from '../pages/Shared/Shared';
-import Conversation from '../../src/shared/Conversation/Conversation';
+import Messages from '../pages/Messages/Messages';
 
 import AuthRoute from './middleware/auth.route';
 import UserRoutes from './middleware/user.route';
@@ -17,7 +17,7 @@ const routes = () => {
             <UserRoutes path="/temporary" exact component={App} />
             <UserRoutes path="/" exact component={Home} />
             <AuthRoute path="/login" exact component={Login} />
-            <UserRoutes path='/conversation/:id' exact component={Conversation} />
+            <UserRoutes path='/messages/:id' exact component={Messages} />
             <UserRoutes path="/shared" exact component={Shared} />
         </Switch>
     );
