--- src/shared/Conversation/component/ConvoHeader.scoped.css
@@ -1,5 +1,0 @@
-*, *::before, *::after {
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;    
-}
